[#77278796] Manage /etc/default/redis-server
This is the stock configuration file. It will be a noop for all existing
machines where the content will match. In a separate commit I'll use this to
raise the nofile ulimit for `logs-redis-N.management`.

Ideally this should be in the upstream module. But a) we need it quickly, b)
it's Debian specific, c) we only need it for as long as we use logship.