[#77278796] Raise ulimit/maxclients for logs-redis
The default nofiles ulimit for services like Redis is 4096. This means that
Redis will configure `maxclients` to 32 less of this value.

We have exceeded this for `logs-redis-N` because we have one logship process
per log file each with its own connection. Double this to account for the
extra machines and applications we have brought up recently.