Use /dev/sdb1 for /dev/mapper/crawler-worker
When I removed the disk from the `mirrorer-1.management` VM in Preview,
the disk that backed /dev/mapper/crawler-worker was demoted from
`/dev/sdc` to `/dev/sdb`.

LVM did the Right Thing and used the UUID to identify the 'physical'
volume:

    mattbostock@preview-mirrorer-1:~$ sudo pvs
    PV         VG      Fmt  Attr PSize  PFree
    /dev/sda2  os      lvm2 a-   49.52g    0
    /dev/sdb1  crawler lvm2 a-   50.00g    0
    mattbostock@preview-mirrorer-1:~$

The purpose of this change is to fix the Hiera data so that Puppet no
longer complains that `/dev/sdc1` does not exist.

This should only be _deployed_ once the extra disk has been removed from
Staging and Production.