[Bouncer] Add businesslink config to serve assets
Redirector continues to serve assets for the Businesslink, Businesslink UK
Welcomes and DirectGov sites which are hotlinked from domains branded with the
branding from those sites. We need to continue to serve these assets, but we
need to turn off Redirector, which means we need to support this behaviour in
Bouncer.

I've copied this from the UKBA proxying we did which ran along similar lines:
e0869ccb45279ece4e6e17474094e2c5059bff44

It is based on the nginx config in Redirector:
https://github.com/alphagov/redirector/blob/1c3edb7e4dad762b9e082a339dcfef0d15286d32/configs/businesslink.conf#L53

If this works, I'll duplicate it for the two other sites it is needed for.