Postgres recovery.conf file for slave
The presence of this file instructs the slave to connect to the master (with
authentication) on start.

We're taking some variables from the upstream postgresql module so that we
don't need to construct them ourselves and they don't get out-of-step. The
tests should flag up any backwards-incompatible changes in the future.

It's worth noting that this will prevent Puppet from successfully
provisioning a new slave because postgres will refuse to start with
`standby_mode=on` and an empty database.