Prevent direct use of govuk_postgresql::server
Introduce a new class called `govuk_postgresql::server::standalone` and use
that in place of `govuk_postgresql::server` for nodes that don't require
replication.

This will allow us to distinguish standalone and master servers (which we
can create roles and databases on) from slaves (where we can't). I'll be
using virtual resources and tags in a following commit to implement that.