Update Vagrantfile for precise box and new trusty
I've changed the default version of the precise box and added trusty
(Ubuntu 14.04) which is still a work in progress as far as Puppet is
concerned but is at least now available for development.

I've also reorganised the Vagrantfile slightly for readability

- virtualbox_box_config() -> default_box_config() and contains all the
  defaults
- Provider specific options are reorganised into one section per
  provider and override the defaults
- Then common config is loaded

The default box version is now 20140828 which has just been uploaded
after testing. The default dist is precise, but can be overridden to
trusty using the `govuk_dev_dist` environment variable.