Record PostgreSQL XLOG position with collectd
For all nodes - standalone, master, and slave. See inline comment for some
details about the query. I'm going to use this for an alert with
`icinga::check::graphite` in a following commit - where I'll explain why
I've taken this approach.

Using collectd's `bytes` type mostly so that we have a better idea of why
unit this is. It's a gauge that is bounded at zero to infinity.

    vagrant@transition-postgresql-slave-1:~$ awk '$1 == "bytes"' /usr/share/collectd/types.db
    bytes                   value:GAUGE:0:U