Alert on PostgreSQL master/slave replication lag
Raise an alert if the difference in XLOG position between the master and
slave exceeds a certain threshold in bytes. This indicates that the slave is
lagging behind, replication has stopped, or one of the nodes is down.

I've plucked the warning and critical thresholds out of thin air because we
don't know enough about our usage yet to predict more accurately. We'll
probably need to tweak these soon - we may even want to align them with
`wal_keep_segments` on the master.

From my testing this appears to be reliable enough. It doesn't suffer from
the same problem as the MySQL lag check where we tried to use
`transformNull()` in 02ca366 because we have two machines independently
reporting values.

We do still need to use `--droplast 1` to prevent `diffSeries()` from
blipping up and effecting the overall average when one node has comitted for
that 10sec timeslot and the other one hasn't (so it compares against NULL).
This doesn't affect most of our other checks because we don't preprocess
them with a Graphite function to remove the NULLs and `check_graphite` then
removes the NULL values from the dataset before comparing.

I considered and discounted a couple of other ways of monitoring this:

- Querying `pg_stat_replication` on the master

  This returns a table of all connected slaves, their state, and log
  positions. This appears to be the closest to checking the slave threads on
  MySQL. It may have been more reliable but:

  - it didn't give us a graph of the replication lag. Which has proved
    useful for MySQL to determine whether it's actually recovering and how
    to tune our thresholds.
  - it requires superuser permissions or a prepared function (applied to the
    master and replicated to the slave) with `SECURITY DEFINER`.
  - it only reports the slave's IP address because it relying on reverse DNS
    to resolve it to a hostname and we don't have any internal DNS. So we'd
    have trouble matching the check back to the slave's FQDN.

- Using `check_postgres --action hot_standby_delay`

  Which fetches the XLOG positions from both the master and slave, in much
  the same way, and compares them. Except:

  - it would require additional HBA rules to allow NRPE on the master or
    slave to connect to the other node.
  - it also wouldn't give us graphs.
  - it didn't seem to cope well with the slave being further ahead than the
    master - reported the negative value as `OK`.
  - not very well documented.

- Using `check_postgres --action replicate_row`

  Which issues an `UPDATE` statement on the master and checks how long it
  takes to replicate to the slave. We could supplement our existing check
  with this in the future, but:

  - also required extra permissions to access both nodes over the network.