Reset net.ipv4.tcp_timestamps to Ubuntu default
This sysctl setting was changed from the default of `1` to `0` as part of
the big "Initial commit" without an explanation of why.

It features in many Linux performance tuning tutorials, also often without a
really good explanation. It seems to boil down to reducing:

- CPU load
- 10byte overhead on each packet
- OS detection from port scanning

However it seems that it can have equally negative effects on fast ethernet
connections. Combined with the fact that most distro defaults are actually
quite sensible, I think we should restore this to the default.

NB: The way that this was previously applied with a file resource meant that
it is inconsistent on some machines because the setting only takes effect
after Puppet has been applied *and* the machine rebooted. By setting it back
with the thias/sysctl module it will take effect immediately.