Add an icinga check for number of redis clients
The upgrade to redis 2.6.6 introduced a bug whereby connections are
not being properly closed, and when we reach the maximum number of
connections the whitehall healthcheck fails.

We think we should need less than 250 connections, and maxclients
is currently set dynamically to 3984. This check warns at 1000 and
is critical at 2000.