Whitehall uses zipinfo from unzip package
This puts in code a previously hidden dependency that Whitehall has on
the system unzip package. When extracting filenames from a zip upload,
we shell out to the `zipinfo` command which is provided by the `unzip`
package.

The `unzip` package was installed for unrelated reasons in acd796ce so
we should ensure that Whitehall requires the package so that if it's
removed we get some warning.