define whitehall-frontend-{6,7}
We want to create a couple more whitehall-frontend machines for
Production. This allows the host to be created and puppet run, prior to
actually putting it in to loadbalancer rotations (which will be done
with hiera in deployment  later)

The reason we are creating these is because rebooting a single machine
(of 5) in Production causes the other 4 to experience high CPU. At the
moment we plan to increase the number of machines to 7, but run 8
workers per machine (rather than 10). This means that when a single
machine is rebooted, our worker count drops from 56 to 48. Currently
rebooting one machine drops our worker count from 50 to 40.

Note that this is only a problem between about 1000hrs and 1600hrs,
after that traffic drops such that rebooting a machine does not send the
rest into overdrive.