[#77965698] Restore redis idle timeout setting
The old module that we were using for Redis 2.4 prior to the 2.6 upgrade in
faa2b70 defaulted to a timeout of 300 seconds for idle connections:

https://github.com/thomasvandoren/puppet-redis/blob/0.10.0/manifests/params.pp#L24

When this was set back to the upstream default of 0 (disabled) it resulted
in connections where the client has abrutly exited without issuing a QUIT
command (and potentially no FIN or RST packet?) remaining in ESTABLISHED
state on the server side and consuming file descriptors + `maxclients`. They
never leave ESTABLISHED state because the server isn't sending any data over
the socket - it's waiting for the client's next command.

These connection counts have been rising steadily since we disabled the
option, presumably due to processes (particulary whitehall-frontend unicorn
workers) dying. It's been worsened recent machine reboots for kernel
upgrades in the past few days - where we see much sudden rises in the number
of connections.

The correct fix is to use the `tcp_keepalive` option which can detect idle
and absent connections. However it's not availble in the version we're
currently using, so restoring the previous `timeout` setting should give us
enough time to review changelogs and produce a new package.