Revert "[#77278796] Raise ulimit/maxclients for logs-redis"
This reverts commit a8a50a5981219d307fbcab19806ef00e3e1dafa9,
4c91fa77c02f790332cc9b9e2aac4216b7a9ecdf, and partially
b390238ff5aee77b53ca45fb6a085fec933dc243.

The rise in connected clients was not legit. It was caused by disabling the
`timeout` setting when we upgraded from redis 2.4 to 2.6 and we're going to
reinstate that setting for the moment.

We can remove this file when the default has been restored on all
logs-redis-N machines.