RabbitMQ partition check requires check_http_timeout_noncrit
The rabbitmq network partition check requires the
`check_http_timeout_noncrit` script which is usually only installed on
the Icinga server.

This factors that out into the icinga::plugin::check_http_timeout_noncrit
class, which is then included in `monitoring::checks` (for the Icinga
server) and `govuk_rabbitmq::monitoring` (for the RabbitMQ servers).