Disable ES GC alerts on support-1 machine
These alerts don't work against this version of Elasticsearch. We've had
them acknowledged in Icinga for about 6 months, so we should just note
that they don't work and disable them in Hiera.

There is another piece of work to address upgrading ES on this machine,
so we can worry about re-enabling this check when that happens.