Check for JSON in the collections API healthcheck.
With the `json_health_check` parameter enabled, we can check not only
that the app is running, but that it self-reports as healthy; if there
are any problems, we can report more detail as to which checks are
failing. This got merged in 2afb8e64ab6ab8b95fd0c331bad4ff8141b327c4.