Add goenv support to govuk_setenv.
This is necessary to allow running Go applications on the dev VM from
foreman in the development repository.  `govuk_setenv` sources
`/etc/environment`, which clobbers the existing PATH.  It's therefore
necessary to add the goenv shims dir back to the path.

I've taken care to ensure that this change is a no-op everywhere except
the dev VM, otherwise the change to govuk_setenv would trigger a restart
of ALL THE THINGS.