Disable attachment text-extraction
This was setup to run as part of the virus scanning process.  It creates
a `.txt` file alongside each attachment with the output from running
tika over the file.  This has proved problematic as some attachments can
cause tika to get wedged.  This blocks the whole virus scanning queue.

The text extraction process adds significantly to the time taken to
process uploaded files (it probably accounts for c. 90% of the time).
Additionally, the text extraction feature is not currently enabled in
Whitehall in production.

We've therefore taken the decision to turn this processing off for the
moment while we decide what to do with this feature.