/etc/puppet/gpg to store GPG keys for Hiera eYAML
Configure Hiera eYAML GPG to expect to find a private GPG key in
`/etc/puppet/gpg` that it can use to decrypt encrypted Hiera keys.

As per Hiera eYAML GPG's documentation[1], the GPG key used by the
Puppet Master (and also `puppet apply`) cannot use a passphrase[2].

While this may sound insecure, in practice I think it's an acceptable
compromise as:

  - the private GPG key will only be stored on the Puppet Master
  - we can store the private GPG key in our encrypted cred store
  - despite not using a passphrase, using a GPG key provides more
    security than we currently have (storing Hiera data in plaintext)

Store the GPG key in `/etc/puppet/gpg` so that we don't need to copy the
key across on each Puppet deployment.

The GPG keys used in Preview, Staging and Production will need to be
copied manually to the Puppet Master when we provision a new
environment. This is preferable to automating this process so that we
can carefully control where the GPG key is stored (in our encrypted cred
store).

[1]: https://github.com/sihil/hiera-eyaml-gpg#configuring-hiera
[2]: https://github.com/sihil/hiera-eyaml-gpg/blob/c2256041558a55f5b84d407707d08c48bb050558/lib/hiera/backend/eyaml/encryptors/gpg.rb#L138-L142