Add safety check for Hiera eYAML GPG
Add a 'safety check' for Hiera eYAML GPG so that we fail early in the
Puppet run if Hiera eYAML GPG is not functioning, the GPG key is not
found or if it fails to decrypt Hiera data correctly.

This check relies on Hiera eYAML GPG being able to successfully decrypt
the key `HIERA_EYAML_GPG_CHECK`. If the key is successfully decrypted,
it will evaluate to `true`; otherwise the `false` default will trigger
the `fail()` condition.

Note that we check for an explicit string with a value of 'true'; it's
not sufficient to check for non-false value as otherwise an encrypted
string would never trigger `fail()`.

This should prevent Puppet from running with missing data, which could
cause a number of problems. One notable example would be modules that
can create users but don't have the capability to change the password of
a user that already exists; if a user was created with an empty
password, we would have no easy way of correcting this.

Note that `development_credentials.yaml` is intentionally not encrypted
so that no GPG keys are required to bring up a development VM.

The safety check in `vagrant_credentials.yaml` is encrypted so that we
can test Hiera eYAML GPG with `vagrant-govuk`. To do this, you'll need
to copy the `vagrant-govuk` GPG keys from the `high` credstore to the
`gpg/` directory in your Puppet repo (on the host machine), e.g:

    ~/govuk/puppet/gpg/