Move pagerduty api key to Hiera
In preparation for changing this key in Production, I have moved it to
Hiera. No environment should be using pagerduty _except_ production, so
I have set the key to be blank in common.yaml

https://github.gds/gds/deployment/pull/406 adds the existing key to
Production

Once this is deployed to Production, I will change the key and redeploy
puppet in Prod.