Send all latest backups offsite
This script previously had a whitelist of hostnames. We have not managed
to keep that list up to date, so were missing a lot of backups in our
offsite set. This changes the script operation, such that it looks for
any MySQL, MongoDB or Postgres latest backup file in any of the
directories in /data/backups, adds the /archived/* backups to that list
and then sends it offsite.

This should mean as we add and remove machines we do not need to update
this embedded list.

/cc @alext