Remove tika and Java from whitehall/asset machines
With the removal of the text-extraction feature in whitehall, there's no
longer a need for tika anywhere, so this removes it.  It also removes
Java where it was only necessary for tika - on the asset servers and
whitehall frontend machines.

Java has been left on the whitehall-backend machines because it's also
on the regular backend machines, and I'm not 100% sure it's not needed
for anythign else.