[#77370232] Allow VPNless access to Carrenza
In order to ensure that we can run Jenkins jobs automatically, we needed to
bypass two-factor authentication on the Carrenza VPN. Credentials could be
supplied using environment variables in Jenkins.

Carrenza went one better and removed the need for us to use the VPN at all
to run Jenkins jobs, so this adds all the support required in order to make
that a pain-free process.

It adds a new hosts entry as the IP we need to hit has no public DNS
associated with it, which was a conscious design decision, and it then
feature-flags that in Hiera.