Update MapIt with latest postcodes
This updates the postcodes in MapIt to the May 2014 version of the ONS
data. As it took me far too long to work out how to do this, I have also
added a README to the `mapit` module which gives a guide to how to do
this in future.

Updating existing servers will be a case of taking them out of rotation
one by one and recreating the database using this latest dump.

This is for story [#78182548]