log app restarts to statsd [CAUSES APP RESTART]
WARNING: Deploying this change would cause _all_ applications to restart
as their upstart script will change.

This change adds a pre-start script to all the GOV.UK applications which
will notify statsd when they restart.

- It's implemented as pre-start because there is no *-restart phase and
  this is between stop and start. It's also because we already use
  post-start and this would conflict or be more complex to implement.
- I've deliberately chosen the simplest possible method of notifying
  statsd (echo to nc)
- I welcome comments on the statsd implementation. I wasn't sure that
  this is the best key to use (it won't tell you which server
  restarted).

This is needed so that we can annotate our Grafana dashboards with
restarts as per story [#78549560]