Add C extension shared objects to Ruby $LOAD_PATH
We already create a shim script at `/usr/bin/puppet` that adds
`/usr/lib/ruby/vendor_ruby` to Puppet's `$LOAD_PATH` to allow us to use
system (i.e. installed from APT) packages for Puppet, Hiera and Facter.

This is necessary because we run puppet using `rbenv`.

This commit adds a further directory within `vendor_ruby` such that we
can load the shared object (`.so`) files used by system Ruby gems that
act as C extensions. Specifically, we are doing this so that the `gpgme`
Ruby gem (installed from the `ruby-gpgme` APT package) can load this
file:

```
/usr/lib/ruby/vendor_ruby/1.9.1/x86_64-linux/gpgme_n.so
```

We use the `Rbconfig::CONFIG` constants which supply the Ruby ABI
version (`1.9.1` currently) and the system architecture to build the
load path.