Use a string for HIERA_EYAML_GPG_CHECK
Use a string for HIERA_EYAML_GPG_CHECK rather than a boolean to avoid
issues whereby an unencrypted value (as is the case for
`development_credentials.yaml`) would be interpreted as a boolean,
whereas a decrypted value (`vagrant_credentials.yaml`) would be output
as a string type.

This made checking for the correct value more complicated than is
necessary. To simplify the logic for this check, and to use a more
realistic example of a value we may encrypt (i.e. some text), use a
string for the check.

I did have "It's all OK folks", but it felt at odds with our team ethos
somehow.