Prefer $host to $server_name
Nginx's embedded variable `$server_name` may contain one or more
space-delimited hostnames.

Instead of using `$server_name` in URI rewrites, use `$host`
which is determined by the URI, the HTTP request `Host` field, or the
server name matching the request[1]. This ensures that we redirect on
the same hostname as was originally requested by the user agent.

[1]: http://nginx.org/en/docs/http/ngx_http_core_module.html#variables