Merge govuk_jenkins into a single class
These classes were split out when we had a master/slave setup for CI. It is
now only used for standalone machines that run deployments, so we can merge
these back into a single class for simplicity.

I've tried to reproduce the important dependencies by adjusting the
`require` param on the package resource.