Add rtyler/jenkins module the Forge
So that we can slowly transition to using it and delete some of our in-house
module. Most notably because it supports installing plugins.

We use a fork of this module for alphagov/ci-puppet. Commits since the fork
all seem to be related to swarm (the master/slave plugin), so I'm going to
ignore that.