Add GitHub Enterprise cert to keystore for Jenkins
Add the self-signed certificate (read from hiera) to the Oracle 7 keystore
which is used by Jenkins. So that Jenkins can talk to the GitHub Enterprise
API for oAuth without getting certificate trust errors.

This is pretty much the same as alphagov/ci-puppet@24f2087 but we now have a
Puppet module that will do the hard part for us.

The path to `cacerts` is hardcoded because I can't find a way to get this
path out of Java programatically. It'd be easier if we were just using
OpenJDK because it'd be `/etc/ssl/certs/java/cacerts`.

NB: The java_ks resource does an `autorequire` on the `file{}` so we don't
need to list it in the `require` param. But we do need the JRE/JDK to be
installed first.