Move Jenkins data back to /var/lib/jenkins
Which is the default location used by the package. It looks like we have
installed Jenkins to `/mnt/jenkins` since the "Initial commit" and later
symlinked from `/var/lib/jenkins` in 8b7fe43.

By restoring the default directory we'll no longer need to do this dance in
the future. These resources will be a noop on new machines until they are
removed.