Replace Jenkins package/server with rtyler/jenkins
Use the third-party module to manage the package and service for us. We
currently manage the repo (our mirror), install Java, and manage the
firewall ourselves, so these are all disabled.

We have to notify the `jenkins::service` class directly when we modify files
that require a service restart because the parent `jenkins` class doesn't
notify the correct things.