Set JENKINS_ARGS on a single line
We have customised `/etc/default/jenkins` to add `-sessionTimeout` in
8661a45. The rtyler/jenkins module provides a way of adding/modifying
individual lines to this file, but it raises an error if a key is found on
more than one line.

This change, when deployed, will allow us to later set `JENKINS_ARGS` using
the `config_hash` param and stop managing `/etc/default/jenkins`. This comes
at the cost of managing slightly more of the arguments in the init script
than we did previously - but it's still better than managing the entire
file.