Move Jenkins home back to /var/lib/jenkins
Which is the default location used by the package. This is necessary because
the `$::jenkins_plugins` fact, which is used by rtyler/jenkins to determine
whether a given plugin is already installed, looks for plugins in the
default location of `~/plugins`. By having our plugins somewhere other than
the user's home directory it attempts to reinstall them on every run.

We have to use an `exec{}` resource to stop the Jenkins service and the
`ssh-agent` daemon, before running the `usermod` command that the `user{}`
resource normally would, otherwise it complains that it can't make the
change while the user is still logged in.

The absolute paths in the `.bashrc` file have been replaced by `~/` to make
them agnostic of such changes in the future.