Remove github_ca_cert used by Jenkins
This was written to a file in 8700952 but the resulting file doesn't appear
to be used anywhere. The referenced story also doesn't contain any
information about why we wanted to add this. So remove it.

I'm going to remove the `File{}` resource because it introduce a bug that I
want to describe in more detail.