Add github-oauth plugin and it's deps to Jenkins
We're going to use this plugin and GitHub Enterprise to secure access to the
Jenkins that serve deploy.{preview,staging,production.alphagov.co.uk in the
same way that we did for alphagov/ci-puppet.

We also need to specify all of it's dependencies because Jenkins won't
install these for us automatically and will fail to load the plugin if they
don't exist. This will upgrade some of them, such as `git`, by a few
versions for existing installs.

I've put the plugin hash inside the manifest rather than hiera because it's
closely related to the other resources in this class. We may want to move
this out or merge it with a hiera lookup if we want to manage all other
plugins like this in the future.

Some preliminary tests show that it is possible to automate this with Groovy
scripts. Although we might choose to do it by hand and document in the
opsmanual for now. We'll see how it goes.