Replace Jenkins .bashrc with .profile
The `keychain` which is started by `.bashrc` stopped being executed after
moving the Jenkins home directory in the previous commit

This is because `/var/lib/jenkins/.profile` didn't exist. This file is
responsible for sourcing `.bashrc` if the shell is bash. It is copied from
`/etc/skel` when human users are created with `useradd -m`, but not
necessarily for system accounts and certainly not when we repointed the home
directory.

Make this work again by removing one layer of indirection and writing to
`.profile` directly. Removing any existing copies of `.bashrc`.