Add role-strategy plugin to Jenkins
As used on alphagov/ci-puppet. It looks like we're going to need the
flexibility of this, rather than using GithubAuthorizationStrategy which
mostly relies on organisation membership.

Some preliminary tests show that it is possible to automate this with Groovy
scripts. Although we might choose to do it by hand and document in the
opsmanual for now. We'll see how it goes.