Use without_content instead of should_not in spec
This makes it slightly more reliable because we're testing that the file is
in the catalog but doesn't contain the text. Rather than the file not being
in the catalog *and* not containing the text. It's a subtle difference, but
the latter is more likely to give false-negatives, for example it would
still pass if the location of the file was changed in the manifest.