Modify PostgreSQL settings on transition-postgresql-master-1
- effective_cache_size: an estimate of how much space is available
for the operating system's disk cache - see http://www.postgresql.org/docs/9.1/static/runtime-config-query.html#GUC-EFFECTIVE-CACHE-SIZE
- shared_buffers: the amount of memory to allocate to postgresql for
caching data - see http://www.postgresql.org/docs/9.1/static/runtime-config-resource.html#GUC-SHARED-BUFFERS

By default these settings are very low for compatibility with old systems
(128MB for effective_cache_size, 24MB for shared_buffers) but they need to
be increased to get reasonable performance on modern dedicated servers.
The values set for both are what we've found to increase performance
in our testing on Preview, and they're in line with general recommendations.

On PostgreSQL 9.1, shared_buffers is dependent on kernel.shmmax, so for now
we need to increase kernel.shmmax so that it is equal to or greater than
the amount of memory which postgresql will request on startup, which is
partly determined by the value of shared_buffers.

When we are on 9.3 we should be able to remove the setting of kernel.shmmax:

    https://wiki.postgresql.org/wiki/What's_new_in_PostgreSQL_9.3#Switch_to_Posix_shared_memory_and_mmap.28.29

We're only modifying these settings on the master for now because we haven't
yet started testing Bouncer's performance when reading from the slave.