Tweak nginx package to install trusty version.
The currently defined version doesn't exist for trusty - only for
precise.  This moves the version definition into hiera so that we can
specify different versions for trusty and precise.  This is a no-op for
precise machines.

This now uses a default of `ensure => 'present'` because we don't
currently have a way of stubbing automatic parameter lookups in the
tests.