Update statsd module to use newer package.
Statsd comes from out PPA, and is not present in Ubuntu.  Setting it to
ensure => latest is therefore safe and ensures the updated version in
the PPA will be picked up.

This also removes  the puppet-managed upstart script as this is now
provided by the package.  The nodejs dependency is also provided by the
package, so it's unnecessary to duplicate it in the module.