Ensure rbenv shims are absent from path
Ensure that rbenv is absent from the system path when making changes to system
gems.

The purpose of the `with_clean_rbenv_environment` method is to ensure that
rbenv does not interfere with changes to system gems.

It did not previously remove the rbenv shims directory from the system path,
this commit fixes that.

Tested on vagrant:

    vagrant@development:~$ govuk_setenv default irb
    irb(main):001:0> # Before
    irb(main):002:0* ENV['PATH'].split(File::PATH_SEPARATOR).reject{|x| x =~ %r{rbenv/versions}}.join(File::PATH_SEPARATOR)
    => "/usr/lib/rbenv/libexec:/usr/lib/goenv/shims:/usr/lib/rbenv/shims:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games"
    irb(main):003:0> # After
    irb(main):004:0* ENV['PATH'].split(File::PATH_SEPARATOR).reject{|x| x =~ %r{rbenv/}}.join(File::PATH_SEPARATOR)
    => "/usr/lib/goenv/shims:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games"
    irb(main):005:0>