Ignore gpg/random_seed
The `random_seed` file is used by GnuPG to save the state of the
system's internal random number generator. The idea being to increase
performance, according to the GnuPG documentation:

https://github.com/gdb/gnupg/blob/fce6c588696051699955eb74382d4dbd34cc7ec9/NEWS#L1144-L1148

There's no need to store this in SCM as it's ephemeral and will change
for each machine running GnuPG that accesses the keys in the `gpg/`
directory.