Also set kernel.shmall on transition-postgresql-master-1
kernel.shmall is the limit on the total amount of shared memory in the
system. Its units are pages, not bytes as for kernel.shmmax.

Its default value in Preview was 524288, which is larger than our set
value for shared_buffers but just under the amount of shared memory
postgresql was actually requesting. Increasing kernel.shmall manually
to exactly the amount of requested memory did allow postgresql to
restart, but for safety this change increases it a little more to
match the value we're setting for kernel.shmmax.

As with kernel.shmmax, we will not need to set kernel.shmall once we
are on PostgreSQL 9.3.