Raise kibana retention to 14 days
    > As a developer I want logs in kibana going back as far as possible
    > so that I can see trends in application behaviour over more than a
    > week.

Last year, a change to how we did json application logging roughly
tripled the size of the logs-elasticsearch indices, and we had to reduce
the retention to 7 days as a result.

Our current usage is that each day's logs average less than 16GB per day.
We keep two copies of each shard, so our effective daily size around 32GB.

When one elasticsearch server is down, shards are redistributed such
that they are spread amongst the remaining two nodes, which means that
each node must have disk space to support a full copy of the entire set
of indices.

Each server has a 250GB partition dedicated to log indices, so it should
be safe to store (250 GB / 16 GB) days of logs - this is just a little
over 15. For a short time before the rotation script runs, we might have
slighly more than the retention limit's worth of logs so if we set the
limit to 15, we will store 14 days of logs which should be safely under
the limit.

This will mean that in normal operation, each server will use ( 16 GB *
2 * 14 / 3) => 149GB of space and in the case of one node failing, the
remaining two nodes will use less than 220GB of space. If a node fails
just before the rotation script runs, we could be storing up to 236GB
per node, which is still within limits.

This commit raises the rotation age limit in days from 8 to 15 for story
[#79784542]