Repurpose govuk::gor ready for reuse
Genericise the `govuk::gor` class so that it can be reused.

Previously, this class was concerned with configuring traffic replay
from GOV.UK cache nodes in Production to Staging.

I'm about to configure traffic replay for the Bouncer application from
Production to Staging (whose traffic does not pass via the cache nodes),
so I'm repurposing `govuk::gor` so that it can be used to configure
traffic reply for both the cache nodes and the Bouncer application.

The Gor configuration specific to the cache nodes is moved to a new
`router::gor` class, which seems appropriate as the cache nodes belong
to the `router` vDC.