Flagging machines as unsafe to reboot
Based on discussion in https://github.com/alphagov/fabric-scripts/pull/55

This includes a class on every machine `govuk::safe_to_reboot` which then
includes another class depending on whether it is safe to reboot or not.

The default is that all classes of machines are safe to reboot.

This is then overridden per-class via `hieradata/class/` based on the
information in:
https://github.gds/pages/gds/opsmanual/2nd-line/rebooting-machines.html