Set alternative for Ruby only on Precise
Wrap the `alternatives` resource for `ruby` in an `if` block such that
it only applies if the Ubuntu LTS release is Precise.

We needed this for Precise because Ruby 4.8 (ABI version 1.8) is part of
the base system, but the Ruby version we wanted `/usr/bin/ruby` to point
to is version 1.9.3 (ABI version 1.9.1).

Trusty has no alternatives available for `ruby` so setting alternatives
is not necessary in Trusty:

    vagrant@transition-postgresql-master-1:~$ update-alternatives --list ruby
    update-alternatives: error: no alternatives for ruby
    vagrant@transition-postgresql-master-1:~$