Specify a default for ::operatingsystemrelease
The previous commit reverted f78c703, which was a workaround for the
fact that `postgresql::globals` from the Puppetlabs module required
certain facter facts to be present, otherwise it would [fail] with the
error:

    No preferred version defined or automatically detected.

[fail]: httpa://github.com/puppetlabs/puppetlabs-postgresql/blob/79c4651240539b511862a2d5c75160169c78552a/manifests/globals.pp#L89-L92

Instead, specify a stub out a default value for
`::operatingsystemrelease`, which was the only fact that was missing to
satisfy the check in `postgresql::globals`.

Because we've removed the code in f78c703, we can now specify the
PostgreSQL version in Hiera. Specifically, in `common.precise.yaml`, so
that we can define the version used by Ubuntu Trusty in `common.yaml`
and delete `common.precise.yaml` once we no longer need to support
Precise.