Move Transition's PostgreSQL db definition to the right app
The postgres transition_production database was temporarily defined
in the transition_postgres app's db manifest for testing on Preview.

We will soon need to start testing Bouncer using this database in other
environments, so this moves the definition of the postgres database into
its rightful place so that it will be created in Staging and Production.
For now transition-postgres will continue to use the database in Preview.