Remove the unused dump user
This reverts commits 53cd561d and 1083639

We have never used this MySQL user, because to create a user which can
both read and restore all databases doesn't give much advantage over the
root user which we are already using.

The user was partially created in #1674 but never fully configured.

To remove the user via puppet would be needlessly complex as I'd need to
first ensure that the govuk_mysql resources were absent, deploy puppet,
then clean up all the other changes, so I propose to deploy this
everywhere and then clean up the extra users.

A related change will be coming later to remove the dump_password stuff
from hiera, but that doesn't block or require this change.