Tell Bouncer's Gor to listen on all interfaces
[Gor] was listening on the loopback interface (`localhost`). This was
copied from the Gor configuration for the GOV.UK cache nodes, where
Varnish receives its traffic over the loopback interface.

This isn't the case for nginx on the `bouncer-?.redirector` machines, so
we tell Gor to listen to port 80 on all network interfaces. This will
pick up the nginx traffic going to the external interface.

I looked at explicitly specifying the IP address for the external
interface, but it's difficult to be sure of what this is. It might be
the `$::ipaddress_eth0` Facter fact, but using that feels too brittle.

Instead, just listen on all interfaces, which I think is fine as we
firewall off any undesired traffic anyway.

[Gor]: https://github.com/buger/gor