Temporarily use Puppetlabs' APT repo directly
We configure the Aptly mirrors in Puppet, but Puppet can't configure
them because it's failing to download the packages it needs from the
Aptly mirror.

Temporarily make all machines use the Puppetlabs APT repository, which
should allow Puppet to configure the new mirror. Once that's done in
Production, this `FIXME` can be reverted.

I've intentionally placed this at the top of the file because I intend
to remove it in a few hours.