Increase PostgreSQL replication setting wal_keep_segments
This value is the minimum number of WAL segments which will be kept,
in case they are needed by a standby machine for replication. For
transition-postgresql-master-1, the current value of 8 covers less
than an hour of our normal imports and very light human use of Transition.

The new value of 64 equates to 1GB of WAL files, and we currently have 58GB
free on transition's master. This new value is higher than currently
needed for postgresql-master-1, but since there is also plenty of disk
space free there and more new apps will soon be using it, it seems
reasonable to apply this change there too.